Groundwork Event Scripting
A bit messy right now, but wanted to get some events up and running so I'm not attaching a million public references to each other. Seems to work well enough for detecting when the skunk is looking while the dog is drinking. May want to change the drink functionality to a coroutine rather than constantly checking for holding down mouse button.